+++ test/mailers/previews/ae_adverse_event_mailer_preview.rb
@@ -13,4 +13,12 @@ class AeAdverseEventMailerPreview < ActionMailer::Preview
     recipient = User.current.first
     AeAdverseEventMailer.sent_for_review(adverse_event, recipient)
   end
+
+  def assigned_to_team
+    adverse_event = AeAdverseEvent.current.joins(:ae_adverse_event_teams).first
+    team = adverse_event.ae_teams.first
+    sender = User.current.first
+    recipient = User.current.second
+    AeAdverseEventMailer.assigned_to_team(sender, adverse_event, team, recipient)
+  end
 end